refactor: rename req4att type file to credential
