+++ packages/types/src/Attestation.ts
@@ -8,7 +8,7 @@
 import type { DidUri } from './DidDetails'
 import type { ICType } from './CType'
 import type { IDelegationNode } from './Delegation'
-import type { ICredential } from './RequestForAttestation'
+import type { ICredential } from './Credential'
 
 export interface IAttestation {
   claimHash: ICredential['rootHash']
