+++ packages/types/src/index.ts
@@ -27,7 +27,7 @@ export * from './Delegation.js'
 export * from './Identity.js'
 export * from './Message.js'
 export * from './Quote.js'
-export * from './RequestForAttestation.js'
+export * from './Credential.js'
 export * from './Terms.js'
 export * from './DidDetails.js'
 export * from './CryptoCallbacks.js'
